# src/main/resources/api/service_openapi.yaml
openapi: 3.0.3
info:
  description: API definition for the CDA
  version: 3.0.0
  title: CDA API
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"

# testing only
#servers:
#  - url: https://cancerdata.dsde-dev.broadinstitute.org
#  - url: https://cda.cda-dev.broadinstitute.org
#  - url: http://localhost:8080
#  - url: http://35.192.60.10:8080
...
paths:
...
  /api/v1/bulk-data:
    get:
      summary: Return all data in CDA
      description: Return all data in CDA
      operationId: bulkData
      tags:
        - query
      parameters:
        - $ref: "#/components/parameters/Table"
        - $ref: "#/components/parameters/IncludeResultsCount"
        - $ref: "#/components/parameters/ResultOffset"
        - $ref: "#/components/parameters/ResultLimit"
      responses:
        200:
          $ref: "#/components/responses/PagedResponse"
  /api/v1/boolean-query:
    post:
      summary: Execute boolean query
      description: |
        Execute a query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: booleanQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"
        - $ref: "#/components/parameters/IncludeResultsCount"
        - $ref: "#/components/parameters/ResultOffset"
        - $ref: "#/components/parameters/ResultLimit"


      requestBody:
        description: The boolean query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/PagedResponse"
  /api/v1/subjects:
    post:
      summary: Execute Subject query
      description: |
        Execute a Subject query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: subjectQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"
        - $ref: "#/components/parameters/IncludeResultsCount"
        - $ref: "#/components/parameters/ResultOffset"
        - $ref: "#/components/parameters/ResultLimit"

      requestBody:
        description: The subject query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/PagedResponse"
  /api/v1/subjects/files:
    post:
      summary: Execute Subject Files query
      description: |
        Execute a Subject Files query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: subjectFilesQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"
        - $ref: "#/components/parameters/IncludeResultsCount"
        - $ref: "#/components/parameters/ResultOffset"
        - $ref: "#/components/parameters/ResultLimit"

      requestBody:
        description: The subject query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/PagedResponse"
  /api/v1/subjects/counts:
    post:
      summary: Execute Subjects Counts query
      description: |
        Execute a Subjects Counts query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: subjectCountsQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"

      requestBody:
        description: The subjects query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/QueryResponse"
  /api/v1/subjects/files/counts:
    post:
      summary: Execute Subjects File Counts query
      description: |
        Execute a Subjects File Counts query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: subjectFileCountsQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"

      requestBody:
        description: The subjects query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/QueryResponse"
  /api/v1/researchsubjects:
    post:
      summary: Execute Research Subject query
      description: |
        Execute a query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: researchSubjectQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"
        - $ref: "#/components/parameters/IncludeResultsCount"
        - $ref: "#/components/parameters/ResultOffset"
        - $ref: "#/components/parameters/ResultLimit"

      requestBody:
        description: The Research Subject query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/PagedResponse"
  /api/v1/researchsubjects/files:
    post:
      summary: Execute ResearchSubject Files query
      description: |
        Execute a ResearchSubject Files query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: researchSubjectFilesQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"
        - $ref: "#/components/parameters/IncludeResultsCount"
        - $ref: "#/components/parameters/ResultOffset"
        - $ref: "#/components/parameters/ResultLimit"

      requestBody:
        description: The research subject query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/PagedResponse"
  /api/v1/researchsubjects/counts:
    post:
      summary: Execute ResearchSubjects Counts query
      description: |
        Execute a ResearchSubjects Counts query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: researchSubjectCountsQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"

      requestBody:
        description: The research subjects query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/QueryResponse"
  /api/v1/researchsubjects/files/counts:
    post:
      summary: Execute ResearchSubjects File Counts query
      description: |
        Execute a ResearchSubjects File Counts query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: researchSubjectFileCountsQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"

      requestBody:
        description: The research subjects query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/QueryResponse"
  /api/v1/specimens:
    post:
      summary: Execute Specimens query
      description: |
        Execute a query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: specimenQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"
        - $ref: "#/components/parameters/IncludeResultsCount"
        - $ref: "#/components/parameters/ResultOffset"
        - $ref: "#/components/parameters/ResultLimit"

      requestBody:
        description: The specimen query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/PagedResponse"
  /api/v1/specimen/files:
    post:
      summary: Execute Specimen Files query
      description: |
        Execute a Specimen Files query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: specimenFilesQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"
        - $ref: "#/components/parameters/IncludeResultsCount"
        - $ref: "#/components/parameters/ResultOffset"
        - $ref: "#/components/parameters/ResultLimit"

      requestBody:
        description: The specimen query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/PagedResponse"
  /api/v1/specimen/counts:
    post:
      summary: Execute Specimen Counts query
      description: |
        Execute a Specimen Counts query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: specimenCountsQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"

      requestBody:
        description: The specimen query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/QueryResponse"
  /api/v1/specimen/files/counts:
    post:
      summary: Execute Specimen File Counts query
      description: |
        Execute a Specimen File Counts query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: specimenFileCountsQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"

      requestBody:
        description: The specimen query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/QueryResponse"
  /api/v1/diagnosis:
    post:
      summary: Execute Diagnosis query
      description: |
        Execute a query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: diagnosisQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"
        - $ref: "#/components/parameters/IncludeResultsCount"
        - $ref: "#/components/parameters/ResultOffset"
        - $ref: "#/components/parameters/ResultLimit"

      requestBody:
        description: The diagnosis query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/PagedResponse"
  /api/v1/diagnosis/counts:
    post:
      summary: Execute Diagnosis Counts query
      description: |
        Execute a Diagnosis Counts query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: diagnosisCountsQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"

      requestBody:
        description: The diagnosis query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/QueryResponse"
  /api/v1/mutations:
    post:
      summary: Execute Mutation query
      description: |
        Execute a query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: mutationQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"
        - $ref: "#/components/parameters/IncludeResultsCount"
        - $ref: "#/components/parameters/ResultOffset"
        - $ref: "#/components/parameters/ResultLimit"

      requestBody:
        description: The mutation query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/PagedResponse"
  /api/v1/mutations/counts:
    post:
      summary: Execute Mutation Counts query
      description: |
        Execute a Mutation Counts query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: mutationCountsQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"

      requestBody:
        description: The mutation query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/QueryResponse"
  /api/v1/treatments:
    post:
      summary: Execute Treatments query
      description: |
        Execute a query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: treatmentsQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"
        - $ref: "#/components/parameters/IncludeResultsCount"
        - $ref: "#/components/parameters/ResultOffset"
        - $ref: "#/components/parameters/ResultLimit"

      requestBody:
        description: The treatments query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/PagedResponse"
  /api/v1/treatments/counts:
    post:
      summary: Execute Treatments Counts query
      description: |
        Execute a Treatments Counts query composed of conditions on columns combined with boolean operators. The
        generated SQL query is returned in the response.
      operationId: treatmentCountsQuery
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"

      requestBody:
        description: The treatment query
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"

      responses:
        200:
          $ref: "#/components/responses/QueryResponse"
...
  /api/v1/global-counts:
    post:
      summary: Returns counts of the DCS
      description: Return GlobalCounts for schema
      operationId: GlobalCounts
      tags:
        - query

      parameters:
        - $ref: "#/components/parameters/DryRun"

      requestBody:
        description: counts
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"
      responses:
        200:
          $ref: "#/components/responses/QueryResponse"
...
components:
  parameters:
...
    IncludeResultsCount:
      in: query
      name: includeCount
      schema:
        type: boolean
        default: false
      description: Show the number of results available for the query
    ResultOffset:
      in: query
      name: offset
      schema:
        type: integer
        default: 0
      description: The number of entries to skip
    ResultLimit:
      in: query
      name: limit
      schema:
        type: integer
        default: 100
      description: The numbers of entries to return per page of data
    Table:
      in: query
      name: table
      schema:
        type: string
      description: tablename
    DryRun:
      in: query
      name: dryRun
      schema:
        type: boolean
        default: false
      description: If true, don't run the query, only generate and return it
...
  responses:
...
    PagedResponse:
      description: query data response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PagedResponseData"
    QueryResponse:
      description: counts data response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/QueryResponseData"
...